Former-commit-id: 3cedfb38d77c060b6b886891c2213b49c1faf52f

+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Commands/Interview/Base/AnswerQuestionCommand.cs
@@ -6,7 +6,7 @@ namespace WB.Core.SharedKernels.DataCollection.Commands.Interview.Base
     {
         public DateTime AnswerTime { get; private set; }
 
-        protected AnswerQuestionCommand(Guid interviewId, Guid userId, Guid questionId, int[] rosterVector, DateTime answerTime)
+        protected AnswerQuestionCommand(Guid interviewId, Guid userId, Guid questionId, decimal[] rosterVector, DateTime answerTime)
             : base(interviewId, userId, questionId, rosterVector)
         {
             this.AnswerTime = answerTime;
