+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/FlagSetToAnswer.cs
@@ -5,7 +5,7 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 {
     public class FlagSetToAnswer : QuestionActiveEvent
     {
-        public FlagSetToAnswer(Guid userId, Guid questionId, int[] propagationVector)
+        public FlagSetToAnswer(Guid userId, Guid questionId, decimal[] propagationVector)
             : base(userId, questionId, propagationVector) {}
     }
 }
