+++ src/Tests/WB.Core.SharedKernels.DataCollection.Tests/InterviewTests/when_propagated_group_is_conditionally_disabled.cs
@@ -68,7 +68,7 @@ namespace WB.Core.SharedKernels.DataCollection.Tests.InterviewTests
         };
 
         Because of = () =>
-           interview.AnswerNumericIntegerQuestion(userId, questionWhichIsForcesPropagationId, new int[] { }, DateTime.Now, 1);
+           interview.AnswerNumericIntegerQuestion(userId, questionWhichIsForcesPropagationId, new decimal[] { }, DateTime.Now, 1);
 
 
         It should_raise_GroupDisabled_event_with_GroupId_equal_to_propagatedGroupId_with_disablement_condition = () =>
